requirements.txt
requests>=2.25.1
websocket-client>=1.0.0
Pillow>=8.0.0